Move the camera body helpers out of the route module
Vercel's build rejected this branch, and the house gate could not see it:

    Type error: Route "app/api/cameras/route.ts" does not match the required types
    of a Next.js Route. "camerasBody" is not a valid Route export field.

A Next.js route module may export ONLY the names Next recognises, and I had exported
`camerasBody` and `__resetCamerasBody` so the test could import them. `npx tsc
--noEmit` passes on that file and so does the whole suite - the route-export contract
is checked only inside `next build`, during "Linting and checking validity of types",
after tsc has already been satisfied.

The helpers move to lib/cameras/body.ts unchanged. The memo, its registry-array
identity key and the edge TTL are all untouched; only where the symbols live has
changed. The route exports `dynamic` and `GET` and nothing else, and the test imports
the lib module - better anyway, since it no longer pulls a route module into a unit
test.

The same defect on app/api/webcams/route.ts reached main via #120 and is fixed in
#121, which also adds tests/unit/route-export-contract.test.ts so the suite catches
this class rather than Vercel catching it after merge.

Test Files  253 passed (253)
     Tests  2135 passed (2135)
